Moved the breadcrumbs to main
The breadcrumbs were on the topbar and now it is on the main above the page content
